backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
rustworkx
firebase-admin
python-multipart
pydantic-settings
pyjwt
python-dotenv
groq
pdfplumber
pytesseract
pdf2image
pillow
